requirements.txt
discord.py
python-dotenv
gspread
oauth2client
python-telegram-bot
supabase
python-dateutil  # For date handling
